+++ src/parameter/filters/index.ts
@@ -9,6 +9,6 @@ export * from './constants';
 export * from './container';
 export * from './build';
 export * from './errors';
-export * from './parse';
+export * from './parser';
 export * from './types';
 export * from './utils';
