+++ src/parameter/pagination/index.ts
@@ -6,5 +6,5 @@
  */
 
 export * from './errors';
-export * from './parse';
+export * from './parser';
 export * from './types';
