+++ src/parameter/sort/index.ts
@@ -8,5 +8,5 @@
 export * from './build';
 export * from './container';
 export * from './errors';
-export * from './parse';
+export * from './parser';
 export * from './types';
